OkExpressions

Documentation for OkExpressions.

OkExpressions.expr_kwarg_int

OkExpressions.gen_expr_startw

OkExpressions.@addmyexpressions_1

OkExpressions.expr_kwarg_int — Method

expr_kwarg_int(keyword) returns regular expression that matches integer after "keyword=". Noted that whitespace is allowed after "=" but NOT allowd before.

Example

expr = expr_kwarg_int("verbosity")
str = "fit!(mach; verbosity= 5)"
match(expr,str).match

# output

"5"

str = "fit!(mach; verbosity = 5)"
match(expr,str)

# output

nothing

OkExpressions.gen_expr_startw — Method

gen_expr_startw(startingword) returns the regular expression that matches what starts with startingword in a new line; it also matches the tailing immediate colon (":") and whitespaces if any. After that, the remaining characters before a newline is stored as a group match of group name myhead that you can call it as shown in the following example:

targetstr = "KEYNOTE:  My Title"
expr = gen_expr_startw("KEYNOTE")
myhead = match(expr, targetstr).captures[end]

# output

"My Title"

or apply it in replace:

replace("KEYNOTE:  My Title", expr => s"!!! note \g<myhead>")

# output

"!!! note My Title"

OkExpressions.@addmyexpressions_1 — Macro

This macro generates variable named expr_startw_... using the expression generator gen_expr_startw. See gen_expr_startw. Julia automatically generate functions and export them

Note

This macro (and also its test) is superfluous and experimental for metaprogramming.